+++ runtime/dolphin/src/assets_config.rs
@@ -184,6 +184,7 @@ impl pallet_manta_pay::Config for Runtime {
     type WeightInfo = weights::pallet_manta_pay::SubstrateWeight<Runtime>;
     type AssetConfig = DolphinAssetConfig;
     type PalletId = MantaPayPalletId;
+    type Suspender = ();
 }
 
 parameter_types! {