+++ runtime/dolphin/src/lib.rs
@@ -129,7 +129,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("dolphin"),
     impl_name: create_runtime_str!("dolphin"),
     authoring_version: 2,
-    spec_version: 4001,
+    spec_version: 4010,
     impl_version: 1,
     apis: RUNTIME_API_VERSIONS,
     transaction_version: 5,